@@ -1,7 +1,7 @@
 package com.steelkiwi.cropiwa.sample.data;
 
 import android.net.Uri;
-import android.support.annotation.DrawableRes;
+import androidx.annotation.DrawableRes;
 
 import com.steelkiwi.cropiwa.sample.App;
 import com.steelkiwi.cropiwa.sample.R;
