@@ -3,6 +3,7 @@ package com.steelkiwi.cropiwa.image;
 import android.graphics.Bitmap;
 import android.graphics.Rect;
 import android.graphics.RectF;
+import android.util.Log;
 
 /**
  * @author yarolegovich
@@ -31,13 +32,32 @@ public class CropArea {
         this.cropRect = cropRect;
     }
 
-    public Bitmap applyCropTo(Bitmap bitmap) {
-        Bitmap immutableCropped = Bitmap.createBitmap(bitmap,
-                findRealCoordinate(bitmap.getWidth(), cropRect.left, imageRect.width()),
-                findRealCoordinate(bitmap.getHeight(), cropRect.top, imageRect.height()),
-                findRealCoordinate(bitmap.getWidth(), cropRect.width(), imageRect.width()),
-                findRealCoordinate(bitmap.getHeight(), cropRect.height(), imageRect.height()));
-        return immutableCropped.copy(immutableCropped.getConfig(), true);
+    Bitmap applyCropTo(Bitmap bitmap) throws IllegalArgumentException {
+            int x = findRealCoordinate(bitmap.getWidth(), cropRect.left, imageRect.width());
+            int y = findRealCoordinate(bitmap.getHeight(), cropRect.top, imageRect.height());
+            int width = findRealCoordinate(bitmap.getWidth(), cropRect.width(), imageRect.width());
+            int height = findRealCoordinate(bitmap.getHeight(), cropRect.height(), imageRect.height());
+
+            if (x < 0) {
+                x = 0;
+            }
+            if (y < 0) {
+                y = 0;
+            }
+            if(x + width >= bitmap.getWidth()) {
+                x = bitmap.getWidth() - width;
+            }
+            if (y + height >= bitmap.getHeight()) {
+                y = bitmap.getHeight() - height;
+            }
+
+            Log.d(this.getClass().getSimpleName(), "x/y/with/height : " + x + "/" + y + "/" + width + "/" + height);
+            Bitmap immutableCropped = Bitmap.createBitmap(bitmap,
+                    x,
+                    y,
+                    width,
+                    height);
+            return immutableCropped.copy(immutableCropped.getConfig(), true);
     }
 
 
